+++ tests/tests_lite/test_lite.py
@@ -22,14 +22,14 @@ import torch.nn.functional
 from tests_lite.helpers.runif import RunIf
 from tests_lite.helpers.utils import no_warning_call
 from torch import nn
-from torch.utils.data import DataLoader, DistributedSampler, Sampler
+from torch.utils.data import DataLoader, DistributedSampler, RandomSampler, Sampler, SequentialSampler, TensorDataset
 
 from lightning_lite.lite import LightningLite
 from lightning_lite.plugins import Precision
 from lightning_lite.strategies import ParallelStrategy, SingleDeviceStrategy, Strategy
 from lightning_lite.utilities import _StrategyType
 from lightning_lite.utilities.exceptions import MisconfigurationException
-from lightning_lite.utilities.seed import pl_worker_init_function
+from lightning_lite.utilities.seed import pl_worker_init_function, seed_everything
 from lightning_lite.utilities.warnings import PossibleUserWarning
 from lightning_lite.wrappers import _LiteDataLoader, _LiteModule, _LiteOptimizer
 
@@ -296,6 +296,38 @@ def test_setup_dataloaders_distributed_sampler_not_needed():
     lite = EmptyLite()
     lite_dataloader = lite.setup_dataloaders(dataloader, replace_sampler=True)
     assert lite_dataloader.sampler is custom_sampler
+
+
+class DistributedSamplerLite(LightningLite):
+    def run(self):
+        # no lite.launch(): pretend we are on rank 0 now
+
+        dataset = TensorDataset(torch.arange(8))
+
+        # shuffling turned off
+        no_shuffle_dataloaders = [
+            DataLoader(dataset),
+            DataLoader(dataset, shuffle=False),
+            DataLoader(dataset, sampler=SequentialSampler(dataset)),
+        ]
+        for dataloader in no_shuffle_dataloaders:
+            dataloader = self.setup_dataloaders(dataloader)
+            expected = [0, 2, 4, 6] if self.global_rank == 0 else [1, 3, 5, 7]
+            assert list(t[0].item() for t in iter(dataloader)) == expected
+
+        # shuffling turned on
+        shuffle_dataloaders = [DataLoader(dataset, shuffle=True), DataLoader(dataset, sampler=RandomSampler(dataset))]
+        for dataloader in shuffle_dataloaders:
+            seed_everything(1)
+            dataloader = self.setup_dataloaders(dataloader)
+            expected = [5, 0, 2, 1] if self.global_rank == 0 else [4, 7, 3, 6]
+            assert list(t[0].item() for t in iter(dataloader)) == expected
+
+
+def test_setup_dataloaders_distributed_sampler_shuffle():
+    """Test that the DataLoader(shuffle=True|False) setting gets carried over correctly into the distributed
+    sampler."""
+    DistributedSamplerLite(accelerator="cpu", strategy="ddp_spawn", devices=2).run()
 
 
 @mock.patch.dict(os.environ, {}, clear=True)