+++ tests/tests_pytorch/core/test_lightning_module.py
@@ -425,6 +425,17 @@ def test_proper_refcount():
     assert sys.getrefcount(torch_module) == sys.getrefcount(lightning_module)
 
 
+def test_lightning_module_scriptable():
+    """Test that the LightningModule is `torch.jit.script`-able.
+
+    Regression test for #15917.
+    """
+    model = BoringModel()
+    trainer = Trainer()
+    model.trainer = trainer
+    torch.jit.script(model)
+
+
 def test_trainer_reference_recursively():
     ensemble = LightningModule()
     inner = LightningModule()